requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.8.2
python-multipart==0.0.9
numpy==1.26.4
opencv-python==4.9.0.80
pytesseract==0.3.10
Pillow==10.3.0
scikit-image==0.22.0
